% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_tf_enhancers_to_target_genes.R
\name{build_gh_pairs_from_peaks}
\alias{build_gh_pairs_from_peaks}
\title{Build GH overlap pairs from peaks (GenomicRanges).}
\usage{
build_gh_pairs_from_peaks(atac_tbl, gh_tbl_std)
}
\description{
Build GH overlap pairs from peaks (GenomicRanges).
}
