% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step1_motif_clustering.R
\name{build_similarity_matrix}
\alias{build_similarity_matrix}
\title{Build an NxN similarity matrix from pairwise comparisons (using Ncor)}
\usage{
build_similarity_matrix(motifs, pw)
}
\arguments{
\item{motifs}{Named list of motifs (defines ordering).}

\item{pw}{data.frame from pairwise_comparisons().}
}
\value{
numeric matrix NxN with diag=1 and symmetric entries = Ncor.
}
\description{
Build an NxN similarity matrix from pairwise comparisons (using Ncor)
}
\keyword{internal}
