% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step2_predict_tf_targets.R
\name{check_module2_links}
\alias{check_module2_links}
\title{Perform sanity checks for predicted Module 2 links}
\usage{
check_module2_links(module2)
}
\arguments{
\item{module2}{Module 2 result list or loaded output list.}
}
\value{
TRUE invisibly when valid.
}
\description{
Perform sanity checks for predicted Module 2 links
}
