% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step1_grn_preprocess.R
\name{clean_hgnc}
\alias{clean_hgnc}
\title{Clean HGNC symbols (drop any whitespace suffix), prints a compact diff.}
\usage{
clean_hgnc(df, label = "rna")
}
\arguments{
\item{df}{data.frame with column HGNC}

\item{label}{tag used in printed message}
}
\value{
data.frame with cleaned HGNC column
}
\description{
Clean HGNC symbols (drop any whitespace suffix), prints a compact diff.
}
