% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step1_motif_clustering.R
\name{cutree_to_k}
\alias{cutree_to_k}
\title{Cut an hclust tree to target K clusters (by searching cut height)}
\usage{
cutree_to_k(hc, target_k)
}
\arguments{
\item{hc}{hclust object.}

\item{target_k}{Integer number of clusters desired.}
}
\value{
Integer vector of cluster labels aligned to hc$order labels.
}
\description{
Cut an hclust tree to target K clusters (by searching cut height)
}
\keyword{internal}
