% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step3_diff_grn_pathway.R
\name{default_diff_grn_pathway_databases}
\alias{default_diff_grn_pathway_databases}
\title{Default Enrichr databases for differential GRN pathway analysis}
\usage{
default_diff_grn_pathway_databases()
}
\value{
Character vector of Enrichr library names.
}
\description{
Default Enrichr databases for differential GRN pathway analysis
}
