% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fp_overview_loader.R
\name{discover_overview_motifs}
\alias{discover_overview_motifs}
\title{Discover motif folders that have "<motif>_overview.txt" in at least one sample}
\usage{
discover_overview_motifs(
  root_dir,
  db_name,
  sample_ids = NULL,
  n_samples = NULL,
  verbose = TRUE
)
}
\arguments{
\item{root_dir}{Path containing per-sample folders.}

\item{db_name}{Motif DB folder name inside each sample.}

\item{sample_ids}{Optional subset of sample folder names. Default: discover all.}

\item{n_samples}{Optional limit on number of samples (after filtering).}

\item{verbose}{Print progress.}
}
\value{
Character vector of motif folder names.
}
\description{
Discover motif folders that have "<motif>_overview.txt" in at least one sample
}
