% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_tf_enhancers_to_target_genes.R
\name{.normalize_gh_cols}
\alias{.normalize_gh_cols}
\title{Normalize GH columns to a standard set.}
\usage{
.normalize_gh_cols(df)
}
\description{
Normalize GH columns to a standard set.
}
