% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{EpiScope}
\alias{EpiScope}
\title{Run EpiScope shiny app}
\usage{
EpiScope(dev = FALSE)
}
\arguments{
\item{dev}{Run the applicaiton in developer mode}
}
\value{
The shiny app will open
}
\description{
Run EpiScope shiny app
}
\examples{
\dontrun{
EpiScope()
}
}
