% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareID.R, R/ExtractAnnotationfromOrgDb.R
\name{ExtractGeneInfo}
\alias{ExtractGeneInfo}
\title{ExtractGeneInfo}
\usage{
ExtractGeneInfo(annoDb)

ExtractGeneInfo(annoDb)
}
\arguments{
\item{annoDb}{AnnotationDB}
}
\value{
data.frame

data.frame
}
\description{
Given an AnnotationDB obtained from AnnotationHub, retrieve gene symbol, ensembl ID and Entrez ID

Given an AnnotationDB obtained from AnnotationHub, retrieve gene symbol, ensembl ID and Entrez ID
}
