% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step1_grn_preprocess.R
\name{fp_manifest_trim}
\alias{fp_manifest_trim}
\title{Trim a single leading "_" and rename files accordingly}
\usage{
fp_manifest_trim(df, rename_files = TRUE, verbose = TRUE)
}
\description{
Trim a single leading "_" and rename files accordingly
}
