% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddGeneAnno_Local_AnnoDb.R, R/Utilities.R
\name{getGeneAnno}
\alias{getGeneAnno}
\title{getGeneAnno}
\usage{
getGeneAnno(annoDb, geneID, type)

getGeneAnno(annoDb, geneID, type)
}
\arguments{
\item{annoDb}{annotation package}

\item{geneID}{query geneID}

\item{type}{gene ID type}
}
\value{
data.frame

data.frame
}
\description{
get gene annotation, symbol, gene name etc.

get gene annotation, symbol, gene name etc.
}
\author{
G Yu

G Yu
}
