% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ggheat.R
\docType{package}
\name{ggheat-package}
\alias{episcope}
\alias{episcope-package}
\alias{ggheat-package}
\title{- New: preserves the unscaled input matrix and makes it available (in heatmap order) as `matrix_ordered`.
- Changed default clustering method to `"ward.D2"`
- Updated default `border_color` to `NA` (no cell borders).
- New: `annotation_row_side` ("left" [default] or "right") to place row annotations.
Pretty heatmaps}
\description{
Implementation of heatmaps that offers more control
over dimensions and appearance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oncologylab/episcope}
  \item Report bugs at \url{https://github.com/oncologylab/episcope/issues}
}

}
\author{
\strong{Maintainer}: Yaoxiang Li \email{yaoxiang@example.com}

Authors:
\itemize{
  \item Chunling Yi \email{chunling@example.com}
}

}
\keyword{internal}
