% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_grn_link_network_topic.R
\name{grn_topic_plot_static}
\alias{grn_topic_plot_static}
\title{Plot a static topic network with ggplot2}
\usage{
grn_topic_plot_static(
  net,
  out_file = NULL,
  width = 7,
  height = 5,
  seed = 123,
  title = NULL,
  use_delta = FALSE
)
}
\arguments{
\item{net}{network list}

\item{out_file}{optional PDF path}

\item{width, height}{inches for PDF}

\item{seed}{reproducible layout}

\item{title}{optional}

\item{use_delta}{logical}
}
\value{
ggplot
}
\description{
Plot a static topic network with ggplot2
}
