% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step1_grn_preprocess.R
\name{init_motif_db}
\alias{init_motif_db}
\title{Load motif database and derive TF list}
\usage{
init_motif_db(db, ref_genome = NULL)
}
\arguments{
\item{db}{Character; supported values: "JASPAR2024", "HOCOMOCOv13".}

\item{ref_genome}{Character; "hg38" or "mm10" (used for JASPAR2024 filenames).}
}
\value{
List with `motif_db` tibble and `tf_list` character vector.
}
\description{
Load motif database and derive TF list
}
