% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_multiomic_object.R
\name{is_multiomic_object}
\alias{is_multiomic_object}
\title{Test whether an object uses the compact multiomic schema}
\usage{
is_multiomic_object(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test whether an object uses the compact multiomic schema
}
