% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step3_topic_models.R
\name{load_differential_links}
\alias{load_differential_links}
\title{Load Module 3 differential links}
\usage{
load_differential_links(path, keep_original = FALSE, n_max_files = Inf)
}
\arguments{
\item{path}{Module 3 differential-link directory, manifest CSV, or filtered
link CSV path(s).}

\item{keep_original}{Keep original columns in addition to standardized
Module 3 columns.}

\item{n_max_files}{Optional maximum number of files to load.}
}
\value{
A data.table of standardized differential links.
}
\description{
Load filtered differential-link files produced by
[module3_prepare_differential_links()]. The input can be a Module 3
differential-link directory, a manifest path, or one or more filtered-link
CSV files.
}
