% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_tf_enhancers_to_target_genes.R
\name{load_genehancer_panc}
\alias{load_genehancer_panc}
\title{Load the pancreas GeneHancer ELITE CSV from inst/extdata and normalize.}
\usage{
load_genehancer_panc(
  path = file.path("inst", "extdata", "GeneHancer_v5.24_elite_panc.csv")
)
}
\description{
Load the pancreas GeneHancer ELITE CSV from inst/extdata and normalize.
}
