% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{load_omics_data}
\alias{load_omics_data}
\title{Load a multi-omic data object from disk}
\usage{
load_omics_data(file, verbose = TRUE)
}
\arguments{
\item{file}{Path to an RDS file produced by save_omics_data().}

\item{verbose}{Emit status messages.}
}
\value{
The loaded multi-omic data list.
}
\description{
Load a multi-omic data object from disk
}
