% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step2_connect_tf_genes.R
\name{make_fp_annotation_corr}
\alias{make_fp_annotation_corr}
\title{Recompute corr_fp_tf_* for fp_annotation using fp_score and RNA}
\usage{
make_fp_annotation_corr(
  grn_set,
  method = c("pearson", "spearman", "kendall"),
  cores = 10L,
  chunk_size = 5000L,
  min_non_na = 3L
)
}
\arguments{
\item{grn_set}{List with fp_annotation, fp_score, and rna.}

\item{method}{Correlation method.}

\item{cores}{Number of cores for mclapply on Unix.}

\item{chunk_size}{Pair chunk size.}

\item{min_non_na}{Minimum paired non-NA values required to test.}
}
\value{
fp_annotation with corr_fp_tf_r/p/p_adj.
}
\description{
Recompute corr_fp_tf_* for fp_annotation using fp_score and RNA
}
