% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step2_connect_tf_genes.R
\name{precompute_hvf_hvg_variance}
\alias{precompute_hvf_hvg_variance}
\title{Precompute footprint and gene variance}
\usage{
precompute_hvf_hvg_variance(grn_set, cores = 1L, chunk_size = 50000L)
}
\arguments{
\item{grn_set}{List containing \code{$fp_score} and \code{$rna}.}

\item{cores}{Number of CPU cores to use.}

\item{chunk_size}{Row chunk size for parallel processing.}
}
\value{
A list with \code{fp_variance} and \code{rna_variance} tables.
}
\description{
Precompute footprint and gene variance
}
