% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step1_motif_clustering.R
\name{pwm_revcomp}
\alias{pwm_revcomp}
\title{Reverse-complement a Position Probability Matrix (PPM; rows A,C,G,T)}
\usage{
pwm_revcomp(pwm)
}
\arguments{
\item{pwm}{4 x L numeric matrix (PPM)}
}
\value{
4 x L numeric matrix reversed and complemented
}
\description{
Reverse-complement a Position Probability Matrix (PPM; rows A,C,G,T)
}
\keyword{internal}
