% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step1_grn_preprocess.R
\name{resolve_motif_db_path}
\alias{resolve_motif_db_path}
\title{Resolve motif DB path by db and reference genome}
\usage{
resolve_motif_db_path(db, ref_genome = NULL)
}
\arguments{
\item{db}{Character; "JASPAR2024" or "HOCOMOCOv13".}

\item{ref_genome}{Character; "hg38" or "mm10".}
}
\value{
Character path to motif DB file.
}
\description{
Resolve motif DB path by db and reference genome
}
