% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tflink.R
\name{tflink_data_dir}
\alias{tflink_data_dir}
\title{Show the current TFLink data directory}
\usage{
tflink_data_dir()
}
\value{
A character string with the resolved TFLink base directory.
}
\description{
Show the current TFLink data directory
}
