% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_multiomic_object.R
\name{validate_multiomic_object}
\alias{validate_multiomic_object}
\title{Validate a compact multiomic object}
\usage{
validate_multiomic_object(omics_data)
}
\arguments{
\item{omics_data}{Object to validate.}
}
\value{
The input object, invisibly.
}
\description{
Validate a compact multiomic object
}
