% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step1_motif_clustering.R
\name{best_pwm_alignment}
\alias{best_pwm_alignment}
\title{Compute best alignment similarity for a pair of motifs (PPMs)}
\usage{
best_pwm_alignment(pwm1, pwm2, min_overlap = 1L)
}
\arguments{
\item{pwm1}{4 x L1 probability matrix (PPM)}

\item{pwm2}{4 x L2 probability matrix (PPM)}

\item{min_overlap}{Integer >=1. Minimum overlap length for alignments.}
}
\value{
Named list with cor, Ncor, Ncor1, Ncor2, w1, w2, w, strand, offset.
}
\description{
Compute best alignment similarity for a pair of motifs (PPMs)
}
