% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step2_predict_tf_targets.R
\name{check_predicted_links}
\alias{check_predicted_links}
\title{Perform sanity check for predicted links for Module 2 diagnostics}
\usage{
check_predicted_links(module2)
}
\arguments{
\item{module2}{Module 2 result list or loaded output list.}
}
\value{
TRUE invisibly when valid.
}
\description{
Perform sanity check for predicted links for Module 2 diagnostics
}
