% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step1_motif_clustering.R
\name{load_motif_pfms}
\alias{load_motif_pfms}
\alias{load_motif_pwms}
\title{Load PFMs for a known motif database (converted to PPMs)}
\usage{
load_motif_pfms(db)

load_motif_pwms(db)
}
\arguments{
\item{db}{Character; supported values: "JASPAR2024", "HOCOMOCOv13".}
}
\value{
Named list of 4 x L probability matrices (PPMs).
}
\description{
Load PFMs for a known motif database (converted to PPMs)
}
