% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step2_steps.R
\name{module2_correlate_tf_targets}
\alias{module2_correlate_tf_targets}
\title{Correlate TF expression with target gene expression}
\usage{
module2_correlate_tf_targets(module2_inputs, n_cores = NULL, verbose = TRUE)
}
\arguments{
\item{module2_inputs}{Output from module2_identify_candidate_links.}

\item{n_cores}{Number of worker cores; NULL uses all available cores.}

\item{verbose}{Emit concise progress messages.}
}
\value{
A TF-target correlation table with pass flags.
}
\description{
Correlate TF expression with target gene expression
}
