% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step2_steps.R
\name{module2_identify_candidate_links}
\alias{module2_identify_candidate_links}
\title{Link TFs to potential target genes based on TFBS-TSS proximity or 3D interaction data}
\usage{
module2_identify_candidate_links(
  multiomic_data,
  predicted_tfbs,
  gene_tss = NULL,
  regulatory_prior = NULL,
  project_config = NULL,
  max_distance_bp = NULL,
  verbose = TRUE
)
}
\arguments{
\item{multiomic_data}{CraftGRN multiomic object.}

\item{predicted_tfbs}{Predicted TFBS table or path from Module 1.}

\item{gene_tss}{Optional gene TSS table or path.}

\item{regulatory_prior}{Optional generic FP-target prior.}

\item{project_config}{Optional project config path or list.}

\item{max_distance_bp}{Maximum signed distance to TSS.}

\item{verbose}{Emit concise progress messages.}
}
\value{
A list of normalized Module 2 inputs used by downstream step functions.
}
\description{
Link TFs to potential target genes based on TFBS-TSS proximity or 3D interaction data
}
