% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_step1_motif_clustering.R
\name{parse_jaspar_pfm_text}
\alias{parse_jaspar_pfm_text}
\title{Parse JASPAR Position Frequency Matrix (PFM; counts) text into PPMs}
\usage{
parse_jaspar_pfm_text(jaspar_text)
}
\arguments{
\item{jaspar_text}{Character scalar: full text with one or more motifs in
JASPAR PFM format.}
}
\value{
Named list. Each element is a 4 x L numeric matrix of probabilities
  with rows A,C,G,T.
}
\description{
Parse JASPAR Position Frequency Matrix (PFM; counts) text into PPMs
}
